Add files via upload
A quick response code or QR code is a type of barcode in the form of a matrix. These are machine-readable optical labels that contain pieces of information about the items attached to them. I had created my own QR code generator using HTML, CSS, and JavaScript. The input can be plain text or a URL.